+++ src/main/java/org/kvkit/mail/SendMailG2.java
@@ -2,7 +2,7 @@
  * To change this template, choose Tools | Templates
  * and open the template in the editor.
  */
-package org.kvkit.action;
+package org.kvkit.mail;
 
 /**
  *
