+++ src/main/java/org/kvkit/mail/SendMailG3.java
@@ -1,4 +1,4 @@
-package org.kvkit.action;
+package org.kvkit.mail;
 
 import java.util.Properties;
